Numerous fixes.
Fixed an issue that would cause any new feeds to not be completely parsed.
Fixed an issue that would cause inaccessable feeds to not be completely parsed when the server came back online.
Fixed timestamp issues. The crawler is now timezone aware.
Added handling for cases where the remote feed does not contain a timezone.